requirements.txt
streamlit>=1.26.0
openpyxl>=3.1.0
python-pptx>=0.6.21
pulp>=2.7.0
